livekit-agent/requirements.txt
livekit-agents>=0.8
livekit-plugins-deepgram>=0.6
livekit-plugins-openai>=0.8
livekit-plugins-elevenlabs>=0.7
httpx>=0.27
openai>=1.30
python-dotenv>=1.0
